+++ src/store/slices/trackedTransactions/trackedTransactionsSlice.ts
@@ -1,0 +1,16 @@
+import { StateCreator } from 'zustand/vanilla';
+import { StoreType, MutatorsIn } from 'store/store.types';
+import { TrackedTransactionsSliceType } from './trackedTransactionsSlice.types';
+
+const initialState: TrackedTransactionsSliceType = {};
+
+function getTrackedTransactionsSlice(): StateCreator<
+  StoreType,
+  MutatorsIn,
+  [],
+  TrackedTransactionsSliceType
+> {
+  return () => initialState;
+}
+
+export const trackedTransactionsSlice = getTrackedTransactionsSlice();