+++ src/store/slices/trackedTransactions/trackedTransactionsSlice.types.ts
@@ -1,0 +1,13 @@
+import {
+  TransactionBatchStatusesEnum,
+  TransactionServerStatusesEnum
+} from 'types';
+import { SignedTransactionType } from 'types/transactions.types';
+
+export interface TrackedTransactionsSliceType {
+  [sessionId: string]: {
+    transactions: SignedTransactionType[];
+    status?: TransactionBatchStatusesEnum | TransactionServerStatusesEnum;
+    errorMessage?: string;
+  };
+}