+++ src/store/store.types.ts
@@ -2,6 +2,8 @@ import { AccountSliceType } from './slices/account/account.types';
 import { ConfigSliceType } from './slices/config/config.types';
 import { LoginInfoSliceType } from './slices/loginInfo/loginInfo.types';
 import { NetworkSliceType } from './slices/network/networkSlice.types';
+import { ToastsSliceState } from './slices/toast/toastSlice.types';
+import { TrackedTransactionsSliceType } from './slices/trackedTransactions/trackedTransactionsSlice.types';
 import { TransactionsSliceType } from './slices/transactions/transacitionsSlice.types';
 
 export type StoreType = {
@@ -10,6 +12,8 @@ export type StoreType = {
   loginInfo: LoginInfoSliceType;
   config: ConfigSliceType;
   transactions: TransactionsSliceType;
+  toasts: ToastsSliceState;
+  trackedTransactions: TrackedTransactionsSliceType;
 };
 
 export type MutatorsIn = [